Moved NameContainsKeywordsPredicate.java to predicate package
