+++ src/main/java/seedu/address/model/util/predicate/NameContainsKeywordsPredicate.java
@@ -1,18 +1,17 @@
-package seedu.address.model.person;
+package seedu.address.model.util.predicate;
 
 import java.util.List;
-import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
+import seedu.address.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code Name} matches any of the keywords given.
  */
-public class NameContainsKeywordsPredicate implements Predicate<Person> {
-    private final List<String> keywords;
+public class NameContainsKeywordsPredicate extends ContainsKeywordsPredicate<Person> {
 
     public NameContainsKeywordsPredicate(List<String> keywords) {
-        this.keywords = keywords;
+        super(keywords);
     }
 
     @Override
